requirements.txt
pytest>=7.0.0
coverage>=6.0.0
psycopg2-binary>=2.9.0
pymongo>=4.0.0
PyJWT>=2.0.0
flask>=2.0.0
requests>=2.26.0 